Resolve Copilot review: lift theme resolution into a single provider
useTheme() previously called useLiveQuery() directly, so every leaf
component that renders themed content (ThemedText, ThemedView, cue
rows, etc.) opened its own Drizzle live-query subscription. Move the
live query into a ThemeProvider mounted once near the app root and
have useTheme()/useThemeToggle() read the resolved scheme from
context instead.
